requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-cloud-storage==2.10.0
google-cloud-vision==3.4.5
google-cloud-videointelligence==2.11.4
requests==2.31.0
Pillow==10.1.0
openai==1.3.7
yt-dlp==2023.11.16